WIP navigation fixes
