+++ lib/get_navigation/src/nav2/router_outlet.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 
 import '../../../get.dart';
+import 'get_navigator.dart';
 import 'get_router_delegate.dart';
 
 class RouterOutlet<TDelegate extends RouterDelegate<T>, T extends Object>
